% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolors.R
\name{coolors}
\alias{coolors}
\title{Extract Coolors and Set Palette}
\usage{
coolors(url)
}
\arguments{
\item{url}{The url in " " of the coolors webpage unique to your desired palette.}
}
\value{
A character string of the unique hexadecimal color codes that correspond to the url.
}
\description{
The below functions are used to extract and set
color palettes that will then be applied with
ggpal().
}
