% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpal.R
\name{ggpal}
\alias{ggpal}
\title{Apply a Custom Palette to ggplot()}
\usage{
ggpal(type = "qualitative", aes = "color", midpoint = 0, ...)
}
\arguments{
\item{type}{A character string noting whether the palette to be used is one of \code{"qualitative"}, \code{"diverging"}, \code{"sequential"}, or \code{"binary"}.}

\item{aes}{A character string indicating whether the aesthetic mapping is for the \code{"color"} or \code{"fill"} of the plot.}

\item{midpoint}{A numerical value noting the midpoint used for a palette of type \code{"diverging"}.}

\item{...}{Additional commands to pass to \verb{ggplot2::scale_*} functions in ggplot2.}
}
\value{
\code{ggpal()} calls one of the following, depending on the palette type and aes selected: \verb{scale_*_manual()}, \verb{scale_*_gradient2()}, \verb{scale_*_gradient()}.
}
\description{
This function applies one of four palettes to a ggplot object that have been
set using set_palette(). set_palette() must be used before ggpal() can be used.
This function is a wrapper for ggplot2 functions that can be used to "manually"
update the color palette of a plot.
}
