% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_theme.R
\name{set_theme}
\alias{set_theme}
\title{Set {coolorrr} Palette for ggplots}
\usage{
set_theme()
}
\value{
The function adds four objects to the global environment for the four unique palette types created by a call to set_palette().
}
\description{
The below function can be called to set some special \code{{coolor}} defaults for ggplots.
}
