% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_palette.R
\name{set_palette}
\alias{set_palette}
\title{Extract Coolors and Set Palette}
\usage{
set_palette(
  qualitative = NULL,
  diverging = NULL,
  sequential = NULL,
  binary = NULL
)
}
\arguments{
\item{qualitative}{Either a vector of hexadecimal color codes or a character string url for a multi-valued qualitative color palette selected at coolors.co.}

\item{diverging}{The same as the above, but where the palette must contain three and only three unique colors to indicate the low, mid, and high end values of the diverging palette.}

\item{sequential}{The same as the above, but where the palette must contain two and only two colors to indicate the low and high end values of the sequential palette.}

\item{binary}{The same as the above, but where the palette must contain two and only two colors for two distinct qualitative values.}
}
\value{
The function adds four objects to the global environment for the four unique palette types created by a call to set_palette().
}
\description{
The below function is used to extract and set
color palettes that will then be applied with
ggpal(). If values are NULL, default palettes
are used.
}
